# frontend/openapi/openapi.json
{
  "components": {
    "schemas": {
      "AppointmentCreateRequest": {
        "properties": {
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "end_date": {
            "format": "date",
            "title": "End Date",
            "type": "string"
          },
          "end_time": {
            "anyOf": [
              {
                "format": "time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Time"
          },
          "is_all_day": {
            "default": false,
            "title": "Is All Day",
            "type": "boolean"
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "maintenance_definition_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Maintenance Definition Id"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "reminder_minutes": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reminder Minutes"
          },
          "start_date": {
            "format": "date",
            "title": "Start Date",
            "type": "string"
          },
          "start_time": {
            "anyOf": [
              {
                "format": "time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Time"
          },
          "title": {
            "maxLength": 500,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": ["title", "start_date", "end_date"],
        "title": "AppointmentCreateRequest",
        "type": "object"
      },
      "AppointmentListFilter": {
        "enum": [
          "upcoming",
          "today",
          "past",
          "cancelled",
          "archived",
          "scheduled",
          "completed"
        ],
        "title": "AppointmentListFilter",
        "type": "string"
      },
      "AppointmentRescheduleRequest": {
        "properties": {
          "end_date": {
            "format": "date",
            "title": "End Date",
            "type": "string"
          },
          "end_time": {
            "anyOf": [
              {
                "format": "time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Time"
          },
          "is_all_day": {
            "title": "Is All Day",
            "type": "boolean"
          },
          "start_date": {
            "format": "date",
            "title": "Start Date",
            "type": "string"
          },
          "start_time": {
            "anyOf": [
              {
                "format": "time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Time"
          }
        },
        "required": ["is_all_day", "start_date", "end_date"],
        "title": "AppointmentRescheduleRequest",
        "type": "object"
      },
      "AppointmentResponse": {
        "properties": {
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "end_date": {
            "format": "date",
            "title": "End Date",
            "type": "string"
          },
          "ends_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ends At"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "is_all_day": {
            "title": "Is All Day",
            "type": "boolean"
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "maintenance_definition_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Maintenance Definition Id"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "reminder_minutes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reminder Minutes"
          },
          "start_date": {
            "format": "date",
            "title": "Start Date",
            "type": "string"
          },
          "starts_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Starts At"
          },
          "status": {
            "$ref": "#/components/schemas/AppointmentStatus"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "notes",
          "location",
          "category",
          "reminder_minutes",
          "maintenance_definition_id",
          "is_all_day",
          "start_date",
          "end_date",
          "starts_at",
          "ends_at",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "AppointmentResponse",
        "type": "object"
      },
      "AppointmentStatus": {
        "enum": ["scheduled", "completed", "cancelled", "archived"],
        "title": "AppointmentStatus",
        "type": "string"
      },
      "AppointmentUpdateRequest": {
        "properties": {
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "maintenance_definition_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Maintenance Definition Id"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "reminder_minutes": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reminder Minutes"
          },
          "title": {
            "anyOf": [
              {
                "maxLength": 500,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "title": "AppointmentUpdateRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": ["loc", "msg", "type"],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Self-hosted personal planning platform API",
    "title": "Planforge",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/appointments": {
      "get": {
        "operationId": "list_appointments_endpoint_api_appointments_get",
        "parameters": [
          {
            "in": "query",
            "name": "filter",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/AppointmentListFilter"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Filter"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/AppointmentStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "search",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AppointmentResponse"
                  },
                  "title": "Response List Appointments Endpoint Api Appointments Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Appointments Endpoint",
        "tags": ["appointments"]
      },
      "post": {
        "operationId": "create_appointment_endpoint_api_appointments_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppointmentCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppointmentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Appointment Endpoint",
        "tags": ["appointments"]
      }
    },
    "/api/appointments/{appointment_id}": {
      "delete": {
        "operationId": "delete_appointment_endpoint_api_appointments__appointment_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "appointment_id",
            "required": true,
            "schema": {
              "title": "Appointment Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Appointment Endpoint",
        "tags": ["appointments"]
      },
      "get": {
        "operationId": "get_appointment_endpoint_api_appointments__appointment_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "appointment_id",
            "required": true,
            "schema": {
              "title": "Appointment Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppointmentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Appointment Endpoint",
        "tags": ["appointments"]
      },
      "patch": {
        "operationId": "update_appointment_endpoint_api_appointments__appointment_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "appointment_id",
            "required": true,
            "schema": {
              "title": "Appointment Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppointmentUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppointmentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Appointment Endpoint",
        "tags": ["appointments"]
      }
    },
    "/api/appointments/{appointment_id}/archive": {
      "post": {
        "operationId": "archive_appointment_endpoint_api_appointments__appointment_id__archive_post",
        "parameters": [
          {
            "in": "path",
            "name": "appointment_id",
            "required": true,
            "schema": {
              "title": "Appointment Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppointmentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Archive Appointment Endpoint",
        "tags": ["appointments"]
      }
    },
    "/api/appointments/{appointment_id}/cancel": {
      "post": {
        "operationId": "cancel_appointment_endpoint_api_appointments__appointment_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "appointment_id",
            "required": true,
            "schema": {
              "title": "Appointment Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppointmentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel Appointment Endpoint",
        "tags": ["appointments"]
      }
    },
    "/api/appointments/{appointment_id}/complete": {
      "post": {
        "operationId": "complete_appointment_endpoint_api_appointments__appointment_id__complete_post",
        "parameters": [
          {
            "in": "path",
            "name": "appointment_id",
            "required": true,
            "schema": {
              "title": "Appointment Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppointmentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Complete Appointment Endpoint",
        "tags": ["appointments"]
      }
    },
    "/api/appointments/{appointment_id}/reopen": {
      "post": {
        "operationId": "reopen_appointment_endpoint_api_appointments__appointment_id__reopen_post",
        "parameters": [
          {
            "in": "path",
            "name": "appointment_id",
            "required": true,
            "schema": {
              "title": "Appointment Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppointmentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reopen Appointment Endpoint",
        "tags": ["appointments"]
      }
    },
    "/api/appointments/{appointment_id}/reschedule": {
      "post": {
        "operationId": "reschedule_appointment_endpoint_api_appointments__appointment_id__reschedule_post",
        "parameters": [
          {
            "in": "path",
            "name": "appointment_id",
            "required": true,
            "schema": {
              "title": "Appointment Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppointmentRescheduleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppointmentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reschedule Appointment Endpoint",
        "tags": ["appointments"]
      }
    },
    "/api/appointments/{appointment_id}/restore": {
      "post": {
        "operationId": "restore_appointment_endpoint_api_appointments__appointment_id__restore_post",
        "parameters": [
          {
            "in": "path",
            "name": "appointment_id",
            "required": true,
            "schema": {
              "title": "Appointment Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppointmentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Restore Appointment Endpoint",
        "tags": ["appointments"]
      }
    },
...
  }
}